refactor code, mettre router dans son propre fichier
