+++ Server/Client/Components/Host/pastilleMission.jsx
@@ -1,15 +1,18 @@
 import React from 'react';
 
-export default class PastilleMission extends React.Component {
+class PastilleMission extends React.Component {
   render() {
-    
     let pastille = (<div className={"circleBase pastilleMission " + this.props.color}>
                       {this.props.teamSize ? this.props.teamSize : '0'}
                     </div>);
-    let currentMissionIndicator; 
+    let currentMissionIndicator;
+
     if(this.props.isCurrentMission){
       currentMissionIndicator = (<div className="triangle"></div>);
     }
+
     return <div className="pastilleContainer" >{pastille}{currentMissionIndicator}</div>;
   }
 }
+
+export default PastilleMission;
