+++ Server/Client/Components/Host/playerCard.jsx
@@ -1,16 +1,18 @@
 import React from 'react';
 
-export default class PlayerCard extends React.Component {
+class PlayerCard extends React.Component {
   render() {
     let nbMissionsBleu = "";
     let nbMissionsRouge = "";
     let isPlayerLeader = "";
+
     if(this.props.player.nbMissionBleu > 0){
       nbMissionsBleu = (<div className="player-mission-bleu">
                           <div className="circleBase mini-pastille blue"></div>
                           {this.props.player.nbMissionBleu} Missions bleu
                         </div>);
     }
+
     if(this.props.player.nbMissionRouge > 0){
       nbMissionsRouge = ( <div className="player-mission-rouge">
                             <div className="circleBase mini-pastille red"></div>
@@ -23,8 +25,7 @@ export default class PlayerCard extends React.Component {
                           Leader
                         </div>);
     }
-            
-            
+
     let content = ( <div className="player-card">
                       <div className="player-name">
                         {this.props.player.playerName}
@@ -40,3 +41,5 @@ export default class PlayerCard extends React.Component {
         </div>);
   }
 }
+
+export default PlayerCard;
