+++ Server/Client/app.jsx
@@ -1,10 +1,7 @@
 import React from 'react';
-import { render } from 'react-dom'
-import { Router, Route, Link, browserHistory } from 'react-router'
 import Portal from './Components/portal.jsx';
-import HostMainPage from './Components/Host/hostMainPage.jsx';
 
-export default class App extends React.Component {
+class App extends React.Component {
   render() {
     return (
       <div>
@@ -13,3 +10,5 @@ export default class App extends React.Component {
     )
   }
 }
+
+export default App;
