--- Server/Client/store.js
@@ -1,12 +1,0 @@
-
-import { createStore, combineReducers } from 'redux';
-import {gameReducer} from './Reducers/gameReducer';
-
-// const reducers = combineReducers({
-//   userState: gameReducer,
-//   widgetState: widgetReducer
-// });
-
-const store = createStore(gameReducer);
-
-module.exports = store;
