+++ Server/Client/store.jsx
@@ -1,0 +1,7 @@
+
+import { createStore, combineReducers } from 'redux';
+import gameReducer from './Reducers/gameReducer.jsx';
+
+const store = createStore(gameReducer);
+
+module.exports = store;
